requirements.txt
requests~=2.31.0
beautifulsoup4~=4.12.0
lxml~=4.9.0
langdetect~=1.0.9
python-dotenv~=1.0.0
tqdm~=4.65.0
